+++ processor/src/errors.rs
@@ -5,7 +5,10 @@ use alloc::{boxed::Box, string::String, sync::Arc, vec::Vec};
 
 use miden_core::{deferred::PrecompileError, program::MIN_STACK_DEPTH};
 use miden_debug_types::{Location, SourceFile, SourceSpan};
-use miden_mast_package::debug_info::{DebugSourceNodeId, PackageDebugInfo};
+use miden_mast_package::{
+    PackageDebugInfoError,
+    debug_info::{DebugSourceNodeId, PackageDebugInfo},
+};
 use miden_utils_diagnostics::{Diagnostic, miette};
 
 use crate::{
@@ -19,6 +22,7 @@ use crate::{
 // EXECUTION ERROR
 // ================================================================================================
 
+#[non_exhaustive]
 #[derive(Debug, thiserror::Error, Diagnostic)]
 pub enum ExecutionError {
     #[error("failed to execute arithmetic circuit evaluation operation: {error}")]
@@ -118,6 +122,8 @@ pub enum ExecutionError {
     ProvingError(String),
     #[error(transparent)]
     HostError(#[from] HostError),
+    #[error(transparent)]
+    PackageDebugInfoError(#[from] PackageDebugInfoError),
 }
 
 impl ExecutionError {